automation-scripts/locust/docker-image/locust-tasks/requirements.txt
locustio==0.9.0
google-cloud-pubsub==0.39.1
Faker==1.0.2